Revert commonization of `LegacySavedStateHandleController`
