+++ lifecycle/lifecycle-viewmodel-savedstate/src/jbMain/kotlin/androidx/lifecycle/AbstractSavedStateViewModelFactory.jb.kt
@@ -16,18 +16,14 @@
 
 package androidx.lifecycle
 
-import androidx.annotation.RestrictTo
 import androidx.core.bundle.Bundle
+import androidx.lifecycle.ViewModelProvider.Companion.VIEW_MODEL_KEY
 import androidx.lifecycle.viewmodel.CreationExtras
-import androidx.savedstate.SavedStateRegistry
 import androidx.savedstate.SavedStateRegistryOwner
 import kotlin.reflect.KClass
 
 public actual abstract class AbstractSavedStateViewModelFactory :
-    ViewModelProvider.OnRequeryFactory,
     ViewModelProvider.Factory {
-
-    private var savedStateRegistry: SavedStateRegistry? = null
     private var lifecycle: Lifecycle? = null
     private var defaultArgs: Bundle? = null
 
@@ -37,7 +33,6 @@ public actual abstract class AbstractSavedStateViewModelFactory :
         owner: SavedStateRegistryOwner,
         defaultArgs: Bundle?
     ) {
-        savedStateRegistry = owner.savedStateRegistry
         lifecycle = owner.lifecycle
         this.defaultArgs = defaultArgs
     }
@@ -45,31 +40,17 @@ public actual abstract class AbstractSavedStateViewModelFactory :
     /**
      * {@inheritDoc}
      *
-     * @throws IllegalStateException if no VIEW_MODEL_KEY provided by ViewModelProvider
+     * @throws IllegalStateException if no [VIEW_MODEL_KEY] provided by [ViewModelProvider]
      */
     public override fun <T : ViewModel> create(
         modelClass: KClass<T>,
         extras: CreationExtras
     ): T {
-        val key = extras[ViewModelProvider.VIEW_MODEL_KEY]
+        val key = extras[VIEW_MODEL_KEY]
             ?: throw IllegalStateException(
                 "VIEW_MODEL_KEY must always be provided by ViewModelProvider"
             )
-        // if a factory constructed in the old way use the old infra to create SavedStateHandle
-        return if (savedStateRegistry != null) {
-            create(key, modelClass)
-        } else {
-            create(key, modelClass, extras.createSavedStateHandle())
-        }
-    }
-
-    @OptIn(ExperimentalStdlibApi::class)
-    private fun <T : ViewModel> create(key: String, modelClass: KClass<T>): T {
-        val controller = LegacySavedStateHandleController
-            .create(savedStateRegistry!!, lifecycle!!, key, defaultArgs)
-        val viewModel = create(key, modelClass, controller.handle)
-        viewModel.addCloseable(LegacySavedStateHandleController.TAG_SAVED_STATE_HANDLE_CONTROLLER, controller)
-        return viewModel
+        return create(key, modelClass, extras.createSavedStateHandle())
     }
 
     protected actual open fun <T : ViewModel> create(
@@ -77,16 +58,4 @@ public actual abstract class AbstractSavedStateViewModelFactory :
         modelClass: KClass<T>,
         handle: SavedStateHandle
     ): T = throw NotImplementedError()
-
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    override fun onRequery(viewModel: ViewModel) {
-        // is need only for legacy path
-        if (savedStateRegistry != null) {
-            LegacySavedStateHandleController.attachHandleIfNeeded(
-                viewModel,
-                savedStateRegistry!!,
-                lifecycle!!
-            )
-        }
-    }
 }
